feat: 게시판 작성 기능 * feat: 게시판 작성 기능

+++ src/main/java/com/oinzo/somoim/controller/dto/BoardRequest.java
@@ -7,8 +7,8 @@ import lombok.*;
 @NoArgsConstructor(access = AccessLevel.PROTECTED)
 public class BoardRequest {
 
-    private int clubId;
-    private int userId;
+    private Long clubId;
+    private Long userId;
     private String category;
     private String title;
     private String content;
